add timeouts to the SSE listener
The old code wouldn't react to network interruptions between the
collector and the devices which are monitored. I assumed that the
network stack would "detect" hung TCP connections, but even when the
TCP-level keepalives are used (which is a host-specific setting), they
are usually set to a very long value. So, when the network was
temporarily interrupted, this is what happened:

- the network was cut,
- the server tried sending some telemetry update to us,
- the server did not get a TCP-level ACK in time, and the connection was
  reset/closed on the server,
- the client just wouldn't know.

Fixing this is "tricky". We could implement a client-side TCP keepalive
and set it to a low value, but considering that our RESTCONF
implementation at the server side was recently changed to send some
dummy data periodically (and "more often than once a minute"), let's
rely on that.

Fix the original issue by timing out if the server doesn't send us
anything for over a minute. Also, while we're at that, implement shorter
timeouts for TCP-level connection opening.

Change-Id: I66f2c2bf41de68ce9b6d399cef68b3f84679527e