handle event-stream comments properly
The standard defines "a comment" to be any line which starts with the
colon (`:`) character; these should be ignored.

Once a comment is processed (or ignored, actually), the server still
sends a blank line which would previously trigger immediate event
dispatch. That's not desirable because it would lead to trying to parse
JSON data from an empty string, which once again throws. Solve that by
only performing the dispatch when some data has been received earlier.

Change-Id: I98da49d825fe2c4741937e12fa29a8068f53db97
Reported-by: Michal Hažlinský <hazlinsky@cesnet.cz>